% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double.R
\name{double}
\alias{double}
\title{A Doubling Function}
\usage{
double(x)
}
\arguments{
\item{x}{what you need twice as much of}
}
\description{
Sometimes you just need twice as much. This function can provide that for any 
numerical vector passed in.
}
\details{
Additional paragraph to make sure we have twice as much verbage.
}
\examples{
double(5)
double(3:10)
double("X") # This is a mistake!
}
\keyword{simple}
